[FIXED JENKINS-32046] Rounding error on aggregated throughput in case of different uri report fix (#57)

+++ src/main/java/hudson/plugins/performance/ThroughputUriReport.java
@@ -13,7 +13,7 @@ public class ThroughputUriReport {
         this.uriReport = uriReport;
     }
 
-    public long get() {
+    public double get() {
         if (uriReport.size() == 0) {
           return 0;
         }
@@ -25,6 +25,6 @@ public class ThroughputUriReport {
         if (duration == 0) {
           return uriReport.size(); // more than zero requests should always take at least some time. If that didn't get logged, this is the most suitable alternative.
         }
-        return (long)(uriReport.size() / ((float)duration / MILLISECONDS_IN_SECOND));
+        return (uriReport.size() / ((double)duration / MILLISECONDS_IN_SECOND));
     }
 }
